backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
Pillow==9.3.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
flake8==3.9.2
flake8-docstrings==1.6.0
